+++ src/ionic-tree-provider.ts
@@ -27,11 +27,13 @@ interface IonicState {
   selectedAndroidDeviceName?: string;
   selectedIOSDeviceName?: string;
   projectDirty?: boolean; // Was there a likely change in the project (ie file saved)
+  syncDone: Array<string>; // Was a cap sync done for a particular platform
   outputIsFocused: boolean; // True if the output window is focused
   channelFocus: boolean; // Whether to focus the output window
   refreshDebugDevices: boolean; // Should we refresh the list of debuggable devices
   remoteLogging: boolean; // Whether remote logging is enabled
   configuration: string; // Build configuration
+  project: string; // Angular project name
   packageManager: PackageManager;
   runIOS: Tip;
   runAndroid: Tip;
@@ -48,6 +50,7 @@ export const ionicState: IonicState = {
   workspace: undefined,
   outputIsFocused: false,
   channelFocus: true,
+  syncDone: [],
   refreshDebugDevices: false,
   remoteLogging: false,
   packageManager: PackageManager.npm,
@@ -56,6 +59,7 @@ export const ionicState: IonicState = {
   runWeb: undefined,
   lastRun: undefined,
   configuration: undefined,
+  project: undefined,
 };
 
 export class IonicTreeProvider implements vscode.TreeDataProvider<Recommendation> {