+++ src/process-packages.ts
@@ -63,6 +63,9 @@ export async function processPackages(
     outdated = context.workspaceState.get(PackageCacheOutdated(project));
     versions = context.workspaceState.get(PackageCacheList(project));
     const changed = packagesModified.toUTCString() != packageModifiedLast;
+    if (changed) {
+      ionicState.syncDone = [];
+    }
     if (changed || !outdated || !versions) {
       await Promise.all([
         getRunOutput(outdatedCommand(project), folder).then((data) => {
@@ -157,27 +160,29 @@ export function reviewPackages(packages: object, project: Project) {
     TipType.Capacitor
   );
 
-  project.setGroup('Project', '', TipType.Ionic);
-  project.add(
-    new Tip('Check for Minor Updates', '', TipType.Dependency)
-      .setAction(updateMinorDependencies, project, packages)
-      .setTooltip('Find minor updates for project dependencies')
-  );
-  project.add(
-    new Tip('Security Audit', '', TipType.Files)
-      .setAction(audit, project)
-      .setTooltip('Analyze dependencies using npm audit for security vulnerabilities')
-  );
-  project.add(
-    new Tip('Statistics', '', TipType.Files)
-      .setAction(analyzeSize, project)
-      .setTooltip('Analyze the built project assets and Javascript bundles')
-  );
-  project.add(
-    new Tip('Export', '', TipType.Media)
-      .setAction(ionicExport, project.projectFolder(), ionicState.context)
-      .setTooltip('Export a markdown file with all project dependencies and plugins')
-  );
+  if (project.isCapacitor) {
+    project.setGroup('Project', '', TipType.Ionic);
+    project.add(
+      new Tip('Check for Minor Updates', '', TipType.Dependency)
+        .setAction(updateMinorDependencies, project, packages)
+        .setTooltip('Find minor updates for project dependencies')
+    );
+    project.add(
+      new Tip('Security Audit', '', TipType.Files)
+        .setAction(audit, project)
+        .setTooltip('Analyze dependencies using npm audit for security vulnerabilities')
+    );
+    project.add(
+      new Tip('Statistics', '', TipType.Files)
+        .setAction(analyzeSize, project)
+        .setTooltip('Analyze the built project assets and Javascript bundles')
+    );
+    project.add(
+      new Tip('Export', '', TipType.Media)
+        .setAction(ionicExport, project.projectFolder(), ionicState.context)
+        .setTooltip('Export a markdown file with all project dependencies and plugins')
+    );
+  }
 }
 
 // List any plugins that use Cordova Hooks as potential issue